:truck: Rename package to lamindb_setup
