+++ lamindb_setup/_set.py
@@ -4,7 +4,7 @@ from typing import Union
 from lamin_logger import logger
 from lnhub_rest.core.storage._add_storage import add_storage as add_storage_hub
 
-from lndb.dev.upath import UPath
+from lamindb_setup.dev.upath import UPath
 
 from ._init_instance import register
 from ._settings import settings
@@ -53,7 +53,7 @@ class set:
         logger.success(f"Set storage {root}")
 
 
-@deprecated("lndb.set.storage()")
+@deprecated("lamindb_setup.set.storage()")
 def set_storage(storage: Union[str, Path, UPath]):
     """Deprecated in favor of `set.storage`."""
     set.storage(storage)
