+++ lamindb_setup/dev/_setup_knowledge.py
@@ -41,7 +41,7 @@ def write_bionty_versions(isettings: InstanceSettings):
 
 
 def load_bionty_versions(isettings: InstanceSettings, display: bool = False):
-    """Write CurrentBiontyVersions to ._lndb.yaml in bionty."""
+    """Write CurrentBiontyVersions to ._lamindb_setup.yaml in bionty."""
     if "bionty" in isettings.schema:
         import bionty as bt
         from bionty.dev._io import write_yaml
@@ -71,4 +71,4 @@ def load_bionty_versions(isettings: InstanceSettings, display: bool = False):
         for entity, db in df_lndb.iterrows():
             lndb_dict[entity] = {}
             lndb_dict[entity][db["database"]] = db["database_v"]
-        write_yaml(lndb_dict, basedir / "._lndb.yaml")
+        write_yaml(lndb_dict, basedir / "._lamindb_setup.yaml")
